djangorestapi/requirements.txt
asgiref==3.4.1
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
bson==0.5.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
dataclasses==0.6
Django==3.2.16
django-cors-headers==3.10.1
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
filelock==3.4.1
hyperlink==21.0.0
idna==3.4
importlib-resources==5.4.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==0.10.0
lxml==4.9.1
mongoengine==0.19.1
packaging==21.3
parsel==1.6.0
Protego==0.2.1
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pymongo==3.10.1
pyOpenSSL==22.1.0
pyparsing==3.0.7
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.6.2
requests==2.15.1
requests-file==1.5.1
Scrapy==2.6.3
service-identity==21.1.0
six==1.14.0
sqlparse==0.2.4
tldextract==3.1.2
Twisted==21.2.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
tzdata==2022.5
urllib3==1.26.12
w3lib==2.0.1
zipp==3.6.0
zope.interface==5.5.0
